UserController에 authenticationService 추가
- 인증한 사용자의 Id와 변경 삭제하려는 user의 id가 같은 경우만 변경 삭제 가능

+++ app/src/main/java/com/codesoom/assignment/controllers/UserController.java
@@ -1,14 +1,21 @@
 package com.codesoom.assignment.controllers;
 
+import com.codesoom.assignment.application.AuthenticationService;
 import com.codesoom.assignment.application.UserService;
 import com.codesoom.assignment.domain.User;
 import com.codesoom.assignment.dto.UserModificationData;
 import com.codesoom.assignment.dto.UserRegistrationData;
 import com.codesoom.assignment.dto.UserResultData;
+import com.codesoom.assignment.errors.ProductNotFoundException;
+import com.codesoom.assignment.errors.UserUnauthorizedException;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AuthorizationServiceException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/users")
@@ -16,8 +23,11 @@ import javax.validation.Valid;
 public class UserController {
     private final UserService userService;
 
-    public UserController(UserService userService) {
+    private final AuthenticationService authenticationService;
+
+    public UserController(UserService userService, AuthenticationService authenticationService) {
         this.userService = userService;
+        this.authenticationService = authenticationService;
     }
 
     @PostMapping
@@ -28,18 +38,32 @@ public class UserController {
     }
 
     @PatchMapping("{id}")
+    @PreAuthorize("isAuthenticated()")
     UserResultData update(
+            @RequestAttribute Long userId,
             @PathVariable Long id,
             @RequestBody @Valid UserModificationData modificationData
     ) {
-        User user = userService.updateUser(id, modificationData);
-        return getUserResultData(user);
+        if (userId.equals(id)) {
+            User user = userService.updateUser(id, modificationData);
+            return getUserResultData(user);
+        } else {
+            throw new UserUnauthorizedException();
+        }
     }
 
     @DeleteMapping("{id}")
+    @PreAuthorize("isAuthenticated()")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    void destroy(@PathVariable Long id) {
-        userService.deleteUser(id);
+    void destroy(
+            @RequestAttribute Long userId,
+            @PathVariable Long id
+    ) {
+        if (userId.equals(id)) {
+            userService.deleteUser(id);
+        } else {
+            throw new UserUnauthorizedException();
+        }
     }
 
     private UserResultData getUserResultData(User user) {
